refactor(module): rename services/session_state/service to services/base.py docs: update STRUCTURE.md
